src/pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "smart_news_organizer"
version = "0.1.0"
description = "Program to organize news, this program makes use of AI to give summaries."
readme = "README.md"
authors = [{name = "Fernando Pujaico Rivera", email = "fernando.pujaico.rivera@gmail.com"}]
maintainers = [{name = "Fernando Pujaico Rivera", email = "fernando.pujaico.rivera@gmail.com"}]
requires-python = ">=3.8"
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
]
license = "MIT"
license-files = ["LICENSE"]
keywords = ["writing", "translate"]
dependencies = [
    "PyQt5",
    "deep-consultation",
    "feedparser",
    "python-dateutil",
    "requests"
]

[project.urls]
"Bug Reports" = "https://github.com/trucomanx/SmartNewsOrganizer/issues"
"Funding" = "https://trucomanx.github.io/en/funding.html"
"Source" = "https://github.com/trucomanx/SmartNewsOrganizer"

[project.scripts]
"smart-news-organizer" = "smart_news_organizer.program:main"

[tool.setuptools]
packages = ["smart_news_organizer", "smart_news_organizer.modules"]

[tool.setuptools.package-data]
"smart_news_organizer" = ["icons/logo.png"]
